# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Watering System API",
      "description": "API documentation for Watering System",
      "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "tags": [
      {
        "name": "sensor",
        "description": "Operations related to sensors"
      },
      {
        "name": "valve",
        "description": "Operations related to valves"
      },
      {
        "name": "plantbed",
        "description": "Operations related to plant beds"
      },
      {
        "name": "reading",
        "description": "Operations related to readings"
      }
    ],
    "paths": {
      "/api/sensor": {
        "get": {
          "tags": [
            "sensor"
          ],
          "summary": "Get all sensors",
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Sensor"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "sensor"
          ],
          "summary": "Create a new sensor",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/SensorInput"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Sensor created successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      },
      "/api/sensor/{sensor_id}": {
        "put": {
          "tags": [
            "sensor"
          ],
          "summary": "Update a sensor",
          "parameters": [
            {
              "name": "sensor_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/SensorInput"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Sensor updated successfully"
            }
          }
        },
        "delete": {
          "tags": [
            "sensor"
          ],
          "summary": "Delete a sensor",
          "parameters": [
            {
              "name": "sensor_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Sensor deleted successfully"
            }
          }
        }
      },
      "/api/valve": {
        "get": {
          "tags": [
            "valve"
          ],
          "summary": "Get all valves",
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Valve"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "valve"
          ],
          "summary": "Create a new valve",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/ValveInput"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Valve created successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      },
      "/api/valve/{valve_id}": {
        "put": {
          "tags": [
            "valve"
          ],
          "summary": "Update a valve",
          "parameters": [
            {
              "name": "valve_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/ValveInput"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Valve updated successfully"
            }
          }
        },
        "delete": {
          "tags": [
            "valve"
          ],
          "summary": "Delete a valve",
          "parameters": [
            {
              "name": "valve_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Valve deleted successfully"
            }
          }
        }
      },
      "/api/plantbed": {
        "get": {
          "tags": [
            "plantbed"
          ],
          "summary": "Get all plant beds",
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PlantBed"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "plantbed"
          ],
          "summary": "Create a new plant bed",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/PlantBedInput"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Plant bed created successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      },
      "/api/plantbed/{plantbed_id}": {
        "put": {
          "tags": [
            "plantbed"
          ],
          "summary": "Update a plant bed",
          "parameters": [
            {
              "name": "plantbed_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/PlantBedInput"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Plant bed updated successfully"
            }
          }
        },
        "delete": {
          "tags": [
            "plantbed"
          ],
          "summary": "Delete a plant bed",
          "parameters": [
            {
              "name": "plantbed_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Plant bed deleted successfully"
            }
          }
        }
      },
      "/api/plantbed/{plantbed_id}/activate": {
        "put": {
          "tags": [
            "plantbed"
          ],
          "summary": "Activate a plant bed",
          "parameters": [
            {
              "name": "plantbed_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Plant bed activated successfully"
            }
          }
        }
      },
      "/api/plantbed/{plantbed_id}/deactivate": {
        "put": {
          "tags": [
            "plantbed"
          ],
          "summary": "Deactivate a plant bed",
          "parameters": [
            {
              "name": "plantbed_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Plant bed deactivated successfully"
            }
          }
        }
      },
      "/api/readings": {
        "get": {
          "tags": [
            "reading"
          ],
          "summary": "Get readings within a timestamp range",
          "parameters": [
            {
              "name": "from",
              "in": "query",
              "description": "Start timestamp (format: YYYY-MM-DD HH:MM:SS)",
              "required": true,
              "type": "string"
            },
            {
              "name": "to",
              "in": "query",
              "description": "End timestamp (format: YYYY-MM-DD HH:MM:SS)",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Reading"
                }
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Sensor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Sensor 1"
          },
          "address": {
            "type": "string",
            "example": "123 Main St."
          },
          "bus": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "SensorInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sensor 1"
          },
          "address": {
            "type": "string",
            "example": "123 Main St."
          },
          "bus": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Valve": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Valve 1"
          },
          "pin": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "ValveInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Valve 1"
          },
          "pin": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "PlantBed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Plant Bed 1"
          },
          "valve_id": {
            "type": "integer",
            "example": 1
          },
          "sensor_id": {
            "type": "integer",
            "example": 1
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "PlantBedInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Plant Bed 1"
          },
          "valve_id": {
            "type": "integer",
            "example": 1
          },
          "sensor_id": {
            "type": "integer",
            "example": 1
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Reading": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-04-25T12:00:00Z"
          },
          "temperature": {
            "type": "number",
            "example": 25.5
          },
          "humidity": {
            "type": "number",
            "example": 60.5
          },
          "sensor_id": {
            "type": "integer",
            "example": 1
          }
        }
      }
    }
  }
  